Move the Splitwise refresh regression test into the context suite
The end-to-end cover for the blank Create Event modal lived in its own
file only because the page and modal suites mock LineItemsContext. The
context suite already renders the real provider against mocked axios, so
it hosts the test next to the reducer behaviour it exercises.
