+++ client/src/contexts/__tests__/LineItemsContext.test.tsx
@@ -2,7 +2,8 @@ import { useQueryClient } from '@tanstack/react-query';
 import { act } from '@testing-library/react';
 import userEvent from '@testing-library/user-event';
 import React from 'react';
-import { mockAxiosInstance, render, screen, waitFor } from '../../utils/test-utils';
+import { mockAxiosInstance, render, screen, waitFor, within } from '../../utils/test-utils';
+import LineItemsToReviewPage from '../../pages/LineItemsToReviewPage';
 import { LineItemInterface, LineItemsProvider, useLineItems, useLineItemsDispatch } from '../LineItemsContext';
 import { AuthProvider } from '../AuthContext';
 
@@ -664,6 +665,133 @@ describe('LineItemsContext', () => {
                 expect(screen.getByTestId('line-items-count')).toHaveTextContent('1');
                 expect(screen.getByTestId('line-item-4')).toHaveTextContent('Different transaction - selected');
             });
+        });
+    });
+
+    // End-to-end cover for the bug the selection-preserving populate fixes: the
+    // Splitwise refresh landing mid-edit used to blank the Create Event modal.
+    describe('Refresh Started By Creating A Splitwise Expense', () => {
+        const splitwiseLineItem = {
+            id: '3',
+            date: 1640995200,
+            payment_method: 'Splitwise',
+            description: 'Alex paid Test transaction 1',
+            responsible_party: 'Alex',
+            amount: -25.00,
+        };
+
+        let serverLineItems: LineItemInterface[];
+        let finishSplitwiseRefresh: () => void;
+
+        beforeEach(() => {
+            serverLineItems = mockLineItems;
+            finishSplitwiseRefresh = () => { };
+
+            mockAxiosInstance.get.mockImplementation((url: string) => {
+                switch (url) {
+                    case 'api/auth/me':
+                        return Promise.resolve({
+                            data: { id: 'user_123', email: 'test@example.com', first_name: 'Test', last_name: 'User' }
+                        });
+                    case 'api/line_items':
+                        return Promise.resolve({ data: { data: serverLineItems } });
+                    case 'api/categories':
+                        return Promise.resolve({ data: { data: [{ id: 'cat_dining', name: 'Dining' }] } });
+                    case 'api/tags':
+                        return Promise.resolve({ data: { data: [] } });
+                    case 'api/splitwise/friends':
+                        return Promise.resolve({ data: { data: [{ id: 7, name: 'Alex' }] } });
+                    case 'api/splitwise/current-user':
+                        return Promise.resolve({ data: { data: { id: 1, name: 'Me' } } });
+                    default:
+                        return Promise.reject(new Error(`Unexpected GET ${url}`));
+                }
+            });
+
+            mockAxiosInstance.post.mockImplementation((url: string) => {
+                switch (url) {
+                    case 'api/event-hints/evaluate':
+                        return Promise.resolve({ data: { data: { suggestion: null } } });
+                    case 'api/splitwise/expenses':
+                        return Promise.resolve({ data: { id: 'splitwise_expense_1' } });
+                    case 'api/refresh/account':
+                        // The refresh runs in the background and finishes whenever
+                        // the test decides to resolve it.
+                        return new Promise((resolve) => {
+                            finishSplitwiseRefresh = () => resolve({ data: { message: 'success' } });
+                        });
+                    default:
+                        return Promise.reject(new Error(`Unexpected POST ${url}`));
+                }
+            });
+        });
+
+        const renderReviewPage = () => renderWithProviders(
+            <LineItemsProvider>
+                <LineItemsToReviewPage />
+            </LineItemsProvider>
+        );
+
+        const selectFirstLineItem = async (user: ReturnType<typeof userEvent.setup>) => {
+            const table = await screen.findByRole('table');
+            await waitFor(() => expect(within(table).getAllByRole('checkbox')).toHaveLength(2));
+            await user.click(within(table).getAllByRole('checkbox')[0]);
+        };
+
+        const createSplitwiseExpense = async (user: ReturnType<typeof userEvent.setup>) => {
+            await user.click(screen.getByRole('button', { name: 'Create Splitwise Expense' }));
+            const splitwiseDialog = await screen.findByRole('dialog');
+            await within(splitwiseDialog).findByText('Alex');
+            await user.click(within(splitwiseDialog).getAllByRole('checkbox')[0]);
+            await user.click(within(splitwiseDialog).getByRole('button', { name: 'Create Expense' }));
+            await waitFor(() => expect(screen.queryByRole('dialog')).not.toBeInTheDocument());
+        };
+
+        const completeSplitwiseRefresh = async () => {
+            serverLineItems = [...mockLineItems, splitwiseLineItem];
+            await act(async () => {
+                finishSplitwiseRefresh();
+            });
+            await screen.findAllByText('Alex paid Test transaction 1');
+        };
+
+        it('the refresh finishing does not reset the open Create Event modal', async () => {
+            const user = userEvent.setup();
+            renderReviewPage();
+
+            await selectFirstLineItem(user);
+            await createSplitwiseExpense(user);
+
+            await user.click(screen.getByRole('button', { name: /Create Event/ }));
+            const eventDialog = await screen.findByRole('dialog');
+            await user.clear(within(eventDialog).getByLabelText('Event Name'));
+            await user.type(within(eventDialog).getByLabelText('Event Name'), 'Dinner with Alex');
+            await user.click(within(eventDialog).getByRole('combobox', { name: /category/i }));
+            await user.click(await screen.findByRole('option', { name: 'Dining' }));
+
+            await completeSplitwiseRefresh();
+
+            // Re-query: a form reset remounts the fields, so the original nodes
+            // would keep their values even after being detached from the document.
+            const dialogAfterRefresh = screen.getByRole('dialog');
+            expect(within(dialogAfterRefresh).getByLabelText('Event Name')).toHaveValue('Dinner with Alex');
+            expect(within(dialogAfterRefresh).getByRole('combobox', { name: /category/i })).toHaveTextContent('Dining');
+            expect(within(dialogAfterRefresh).getByText('$50.00')).toBeInTheDocument();
+            expect(within(dialogAfterRefresh).getByRole('button', { name: 'Create Event' })).toBeEnabled();
+        });
+
+        it('the refresh finishing keeps the line item selection', async () => {
+            const user = userEvent.setup();
+            renderReviewPage();
+
+            await selectFirstLineItem(user);
+            await createSplitwiseExpense(user);
+            await completeSplitwiseRefresh();
+
+            const checkboxes = within(screen.getByRole('table')).getAllByRole('checkbox');
+            expect(checkboxes[0]).toBeChecked();
+            expect(checkboxes[1]).not.toBeChecked();
+            expect(checkboxes[2]).not.toBeChecked();
         });
     });
 });
